backend/requirements.txt
alembic==1.16.5
altair==5.5.0
attrs==25.4.0
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-speech==1.47.0
azure-common==1.1.28
azure-core==1.38.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
cmake==4.2.1
colorama==0.4.6
dlib @ file:///C:/Users/varsha%20b%20s/Downloads/dlib-19.22.99-cp39-cp39-win_amd64.whl#sha256=0fd828405c77af2df1ff6a09964a8ba7f14838c538d9017046804f353e9b4bc2
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.2.4
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
mysql-connector-python==8.2.0
narwhals==2.15.0
numpy==1.23.5
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==25.0
pandas==2.1.1
pillow==11.3.0
protobuf==4.21.12
pyarrow==21.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.46
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tomli==2.4.0
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
Werkzeug==2.3.7
zipp==3.23.0
